pyproject.toml
[tool.poetry]
name = "Free-Watch"
version = "0.1.0"
description = "Lite IoT servers as plugin/standalone application"
authors = ["Aradhya-Tripathi <aradhyatripathi51@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.8"
djangorestframework = "^3.12.4"
python-dotenv = "^0.19.0"
pymongo = "^3.12.0"
gunicorn = "^20.1.0"
psutil = "^5.8.0"
black = "^21.9b0"
schema = "^0.7.4"
dnspython = "^2.1.0"
sentry-sdk = "^1.4.3"
PyJWT = "^2.2.0"
channels = "^3.0.4"
channels-redis = "^3.3.0"
uvicorn = "^0.15.0"
websockets = "^10.0"
httpx = "^0.20.0"
rich = "^10.16.0"
requests = "^2.26.0"
keydb = "^0.0.1"

[tool.poetry.dev-dependencies]
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"